fixed pull.sh with carraige return issue

+++ tcpautodump/pull.sh
@@ -4,6 +4,7 @@
 
 files=`adb shell "ls /sdcard/tcpautodump_*"`
 
+device=`adb shell getprop ro.product.model | sed 's/ /_/g'`
 
 arr=($files)
 
@@ -14,18 +15,30 @@ fi
 
 file1=${arr[0]}
 
-filename=$(basename -- "$file1")
+CLEAN=${device// /_}
 
+CLEAN=${CLEAN//[^a-zA-Z0-9_]/}
+
+
+filename=$(basename -- "$file1")
 extension="${filename##*.}"
-output_dir="${filename%.*}" 
+output_dir="${filename%.*}_${CLEAN}" 
 
     
 mkdir "$output_dir"
-
+count=0
 for file in $files; do
-    adb pull "$file" "$output_dir"
-    adb shell "rm -f $file"
+    file=${file//[^a-zA-Z0-9_\/\.\-]/}
+    filename=$(basename -- "$file")
+    adb pull "$file" "${output_dir}/${CLEAN}_${filename}"
+    if [ $? -eq 0 ]; then
+        let count++
+        adb shell "rm -f $file"
+    else
+        echo "failed to pull file: ${file}"
+        echo " skipping... the file will not be removed from the device"
+    fi
 done
 
 echo ""
-echo "${#arr[@]} pcap files have been placed in ./$output_dir/"
+echo "${count} pcap files have been placed in ./$output_dir/"
